requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
pyinstaller>=5.0
Pillow>=9.0
